Merged lockdown and testing with config update
In this commit the policy class has modified the lockdown trigger parameter to be lockdown on and off parameters, the number of tests was also updated to be more dynamic (containing the newly quarantined population, total quarantined population and the waitlist for tests). The triggers are now based on predicted percentage of population that is infected based on the quarantined population.  The update also includes an improved testing system to prevent people with colds only to be able to get tested once (even if they get infected with COVID later).