Allow metric functions to return multiple values.
Allow metric functions used in tabular mode to return multiple
values. Ensure that the return types of each function are the
same/compatible with each other. Raise errors if the return values
have different sizes or have incompatible types. The second argument
can be used to return errors on the values, for example.