Update the instructions to encourage use of virtual environments.
Update the README with instructions on how to configure and use
virtual environments for the code. Clean up some of the writing and
clarify a few things such as the specific Python version required, the
fact that the configure script only needs to be run once, and the fact
that Jupyter can be installed but is optional.